feat: Add 'Deshabilitados' option to FilterSelectComponent
This commit adds the 'Deshabilitados' option to the `FilterSelectComponent` in order to allow users to filter by disabled clients. The `onValueChange` method has been updated to emit the selected value, and the `filterChange` method has been modified to handle the new option. The `disabled` property has been introduced to track the selection of the 'Deshabilitados' option. This change improves the functionality of the component and enhances the filtering capabilities.